suppressing too much output in mkdocs

+++ scripts/build.sh
@@ -10,18 +10,16 @@ if [ -z $1 ]; then
 fi
 docsPath=$1
 
-# if [ $2 == "true" ]; then
-#     mkdocs build 2> /dev/null # /app/docs => /app/site
-# else
-#     mkdocs build
-# fi
-#
-# if [ $? -eq 1 ]; then
-#     log "MkDocs failed to build website"
-#     exit 1
-# fi
+if [ $2 == "true" ]; then
+	mkdocs build 2>/dev/null # /app/docs => /app/site
+else
+	mkdocs build
+fi
 
-mkdocs build
+if [ $? -eq 1 ]; then
+	log "MkDocs failed to build website"
+	exit 1
+fi
 
 rm -rf /app/docs
